backend/requirements.txt
django
djangorestframework
django-cors-headers
tensorflow
pillow
